Move png and jpeg version functions to gd.c so the external library build doesn't include anything from libgd/.
